# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CariKos — Platform Perbandingan Kos API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FasilitasItem {
  id: number;
  nama: string;
  icon: string;
}

export interface KosSummary {
  id: number;
  ownerId: number;
  nama: string;
  alamat: string;
  harga: number;
  /** @nullable */
  deskripsi?: string | null;
  /** @nullable */
  jamOperasional?: string | null;
  /** @nullable */
  aturan?: string | null;
  /** @nullable */
  lat?: string | null;
  /** @nullable */
  lng?: string | null;
  /** @nullable */
  noWa?: string | null;
  createdAt: string;
  /** @nullable */
  avgRating: number | null;
  reviewCount: number;
  fasilitas: FasilitasItem[];
}

export interface ReviewItem {
  id: number;
  kosId: number;
  userId: number;
  rating: number;
  /** @nullable */
  komentar?: string | null;
  /** @nullable */
  foto?: string | null;
  tanggal: string;
  createdAt: string;
}

export interface KosDetail {
  id: number;
  ownerId: number;
  nama: string;
  alamat: string;
  harga: number;
  /** @nullable */
  deskripsi?: string | null;
  /** @nullable */
  jamOperasional?: string | null;
  /** @nullable */
  aturan?: string | null;
  /** @nullable */
  lat?: string | null;
  /** @nullable */
  lng?: string | null;
  /** @nullable */
  noWa?: string | null;
  createdAt: string;
  /** @nullable */
  avgRating: number | null;
  reviewCount: number;
  fasilitas: FasilitasItem[];
  reviews: ReviewItem[];
}

export interface ErrorResponse {
  error: string;
}

export type ListKosParams = {
/**
 * Filter berdasarkan kota (case-insensitive substring)
 */
kota?: string;
/**
 * Filter berdasarkan nama universitas (substring match di alamat)
 */
universitas?: string;
/**
 * Harga minimum per bulan
 */
budget_min?: number;
/**
 * Harga maksimum per bulan
 */
budget_max?: number;
limit?: number;
};

export type ListKosTrendingParams = {
limit?: number;
};

export type ListKosBestValueParams = {
limit?: number;
};

export type ListKosBudgetFriendlyParams = {
limit?: number;
};

export type ListKosHighestRatedParams = {
limit?: number;
};

export type ListKosNewestParams = {
limit?: number;
};