# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CariKos — Platform Perbandingan Kos API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: kos
    description: Kos operations
  - name: fasilitas
    description: Fasilitas operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /kos:
    get:
      operationId: listKos
      tags: [kos]
      summary: List semua kos dengan filter opsional
      parameters:
        - name: kota
          in: query
          schema: { type: string }
          description: Filter berdasarkan kota (case-insensitive substring)
        - name: universitas
          in: query
          schema: { type: string }
          description: Filter berdasarkan nama universitas (substring match di alamat)
        - name: budget_min
          in: query
          schema: { type: integer }
          description: Harga minimum per bulan
        - name: budget_max
          in: query
          schema: { type: integer }
          description: Harga maksimum per bulan
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Daftar kos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KosSummary"

  /kos/trending:
    get:
      operationId: listKosTrending
      tags: [kos]
      summary: Kos trending berdasarkan jumlah review terbanyak
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Kos trending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KosSummary"

  /kos/best-value:
    get:
      operationId: listKosBestValue
      tags: [kos]
      summary: Kos value terbaik (rating tinggi dengan harga terjangkau)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Kos value terbaik
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KosSummary"

  /kos/budget-friendly:
    get:
      operationId: listKosBudgetFriendly
      tags: [kos]
      summary: Kos budget terbaik (harga termurah)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Kos budget terjangkau
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KosSummary"

  /kos/highest-rated:
    get:
      operationId: listKosHighestRated
      tags: [kos]
      summary: Kos dengan rating tertinggi
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Kos rating tertinggi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KosSummary"

  /kos/newest:
    get:
      operationId: listKosNewest
      tags: [kos]
      summary: Kos yang baru ditambahkan
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Kos terbaru
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KosSummary"

  /kos/{id}:
    get:
      operationId: getKos
      tags: [kos]
      summary: Detail lengkap satu kos beserta fasilitas dan review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detail kos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KosDetail"
        "404":
          description: Kos tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /fasilitas:
    get:
      operationId: listFasilitas
      tags: [fasilitas]
      summary: Daftar semua fasilitas yang tersedia
      responses:
        "200":
          description: Daftar fasilitas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FasilitasItem"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    KosSummary:
      type: object
      required:
        - id
        - ownerId
        - nama
        - alamat
        - harga
        - createdAt
        - avgRating
        - reviewCount
        - fasilitas
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        nama:
          type: string
        alamat:
          type: string
        harga:
          type: integer
        deskripsi:
          type: ["string", "null"]
        jamOperasional:
          type: ["string", "null"]
        aturan:
          type: ["string", "null"]
        lat:
          type: ["string", "null"]
        lng:
          type: ["string", "null"]
        noWa:
          type: ["string", "null"]
        createdAt:
          type: string
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        fasilitas:
          type: array
          items:
            $ref: "#/components/schemas/FasilitasItem"

    KosDetail:
      type: object
      required:
        - id
        - ownerId
        - nama
        - alamat
        - harga
        - createdAt
        - avgRating
        - reviewCount
        - fasilitas
        - reviews
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        nama:
          type: string
        alamat:
          type: string
        harga:
          type: integer
        deskripsi:
          type: ["string", "null"]
        jamOperasional:
          type: ["string", "null"]
        aturan:
          type: ["string", "null"]
        lat:
          type: ["string", "null"]
        lng:
          type: ["string", "null"]
        noWa:
          type: ["string", "null"]
        createdAt:
          type: string
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        fasilitas:
          type: array
          items:
            $ref: "#/components/schemas/FasilitasItem"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewItem"

    FasilitasItem:
      type: object
      required: [id, nama, icon]
      properties:
        id:
          type: integer
        nama:
          type: string
        icon:
          type: string

    ReviewItem:
      type: object
      required: [id, kosId, userId, rating, tanggal, createdAt]
      properties:
        id:
          type: integer
        kosId:
          type: integer
        userId:
          type: integer
        rating:
          type: integer
        komentar:
          type: ["string", "null"]
        foto:
          type: ["string", "null"]
        tanggal:
          type: string
        createdAt:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string