# lib/api-zod/src/generated/types/fasilitasItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CariKos — Platform Perbandingan Kos API
 * OpenAPI spec version: 0.1.0
 */

export interface FasilitasItem {
  id: number;
  nama: string;
  icon: string;
}