# lib/api-zod/src/generated/types/kosSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CariKos — Platform Perbandingan Kos API
 * OpenAPI spec version: 0.1.0
 */
import type { FasilitasItem } from './fasilitasItem';

export interface KosSummary {
  id: number;
  ownerId: number;
  nama: string;
  alamat: string;
  harga: number;
  /** @nullable */
  deskripsi?: string | null;
  /** @nullable */
  jamOperasional?: string | null;
  /** @nullable */
  aturan?: string | null;
  /** @nullable */
  lat?: string | null;
  /** @nullable */
  lng?: string | null;
  /** @nullable */
  noWa?: string | null;
  createdAt: string;
  /** @nullable */
  avgRating: number | null;
  reviewCount: number;
  fasilitas: FasilitasItem[];
}