# lib/api-zod/src/generated/types/reviewItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CariKos — Platform Perbandingan Kos API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewItem {
  id: number;
  kosId: number;
  userId: number;
  rating: number;
  /** @nullable */
  komentar?: string | null;
  /** @nullable */
  foto?: string | null;
  tanggal: string;
  createdAt: string;
}